chore(models): небольшое изменение paymentDT под Unix (int -> int64)

+++ models/order.go
@@ -35,7 +35,7 @@ type Payment struct {
 	Currency     string `json:"currency"`
 	Provider     string `json:"provider"`
 	Amount       int    `json:"amount"`
-	PaymentDT    int    `json:"payment_dt"`
+	PaymentDT    int64  `json:"payment_dt"`
 	Bank         string `json:"bank"`
 	DeliveryCost int    `json:"delivery_cost"`
 	GoodsTotal   int    `json:"goods_total"`
